Combine pending-tx-details component into pending-tx-details
These were only separated originally so we could make the notification-based TX approval work, which provided its own buttons.

This two templates are logically highly coupled, and keeping them working while separate has been difficult at times, and has even required resorting to dubious practices, like using React's `refs` pattern.

This combines them into one fairly large component, but I think it's ok, we can still break this up into components, just not the separation that it had previously.
